Add promoted.js file
